requirements.txt
# Nifty Sehwag Strategy - Python Dependencies

# Core dependencies
openalgo>=1.0.0              # OpenAlgo Python client
pyyaml>=6.0.1                # YAML config parsing
pytz>=2023.3                 # Timezone support
pandas>=2.0.3                # Data analysis (candles)
sqlalchemy>=2.0.23           # Database ORM

# Optional (for WebSocket support)
websockets>=12.0             # WebSocket client

# Standard library (included with Python)
# - logging, threading, datetime, typing, pathlib
# - json, uuid, time, sys, os
